Reduced getShippingMethods()
Removed code that was manipulating the cart and setting shipping methods by on customer addresses. 

When using ApplePay from the cart shipping methods are retrieved by ApplePay through the /adyen/applePay/retrieveShippingMethods AJAX request so there is no need to set them here. 

If the customer is logged & in Magento's checkout then the shipping amount is selected in the checkout & sent to ApplePay as a line item. So again no need to retrieve & set shipping methods here.

The code here was changing the quote & interfering with the shipping/delivery estimate function of the cart.